+++ src/main/java/com/driver/services/DriverService.java
@@ -5,6 +5,6 @@ import com.driver.model.Driver;
 public interface DriverService {
 
 		public void register(String mobile, String password);
-		public void removeDriver(Long driverId);
-		public void updateStatus(Long driverId);
+		public void removeDriver(Integer driverId);
+		public void updateStatus(Integer driverId);
 }
